# add benchmark test for endpoint labels