Without metadata:
```
BenchmarkEndpointGeneration/1/100-64                3204            361476 ns/op                 5.79 kb/msg           100 resources/msg           94302 B/op       1511 allocs/op
BenchmarkEndpointGeneration/10/10-64                9411            126428 ns/op                 2.83 kb/msg            10.0 resources/msg         12614 B/op        158 allocs/op
BenchmarkEndpointGeneration/100/10-64               1177            995654 ns/op                26.2 kb/msg             10.0 resources/msg         45132 B/op        164 allocs/op
BenchmarkEndpointGeneration/1000/1-64               1228            956593 ns/op                26.6 kb/msg              1.00 resources/msg        36849 B/op         25 allocs/op
```

With metadata:
```
BenchmarkEndpointGeneration/1/100-64                3214            377944 ns/op                 5.79 kb/msg           100 resources/msg           94344 B/op       1511 allocs/op
BenchmarkEndpointGeneration/10/10-64                9452            129277 ns/op                 2.83 kb/msg            10.0 resources/msg         12627 B/op        158 allocs/op
BenchmarkEndpointGeneration/100/10-64                967           1051356 ns/op                26.2 kb/msg             10.0 resources/msg         46548 B/op        185 allocs/op
BenchmarkEndpointGeneration/1000/1-64               1210            975391 ns/op                26.6 kb/msg              1.00 resources/msg        37935 B/op         41 allocs/op
```